modify autoIDs
