+++ drupal/web/sites/default/files/sitemap.xml
@@ -30,7 +30,7 @@
   <xhtml:link rel="alternate" hreflang="en" href="https://frontend.lndo.site/en/articles/benefits-using-drupal-your-website"/>
   <xhtml:link rel="alternate" hreflang="fi" href="https://frontend.lndo.site/fi/articles/drupalin-hyodyt-verkkosivustollesi"/>
   <xhtml:link rel="alternate" hreflang="sv" href="https://frontend.lndo.site/sv/articles/fordelarna-med-att-anvanda-drupal-din-webbplats"/>
-  <lastmod>2023-12-12T08:40:05+00:00</lastmod>
+  <lastmod>2023-12-14T09:23:30+00:00</lastmod>
   <priority>0.5</priority>
  </url>
  <url>
@@ -38,7 +38,7 @@
   <xhtml:link rel="alternate" hreflang="en" href="https://frontend.lndo.site/en/articles/benefits-using-drupal-your-website"/>
   <xhtml:link rel="alternate" hreflang="fi" href="https://frontend.lndo.site/fi/articles/drupalin-hyodyt-verkkosivustollesi"/>
   <xhtml:link rel="alternate" hreflang="sv" href="https://frontend.lndo.site/sv/articles/fordelarna-med-att-anvanda-drupal-din-webbplats"/>
-  <lastmod>2023-12-12T08:40:05+00:00</lastmod>
+  <lastmod>2023-12-14T09:23:30+00:00</lastmod>
   <priority>0.5</priority>
  </url>
  <url>
@@ -46,7 +46,7 @@
   <xhtml:link rel="alternate" hreflang="en" href="https://frontend.lndo.site/en/articles/benefits-using-drupal-your-website"/>
   <xhtml:link rel="alternate" hreflang="fi" href="https://frontend.lndo.site/fi/articles/drupalin-hyodyt-verkkosivustollesi"/>
   <xhtml:link rel="alternate" hreflang="sv" href="https://frontend.lndo.site/sv/articles/fordelarna-med-att-anvanda-drupal-din-webbplats"/>
-  <lastmod>2023-12-12T08:40:05+00:00</lastmod>
+  <lastmod>2023-12-14T09:23:30+00:00</lastmod>
   <priority>0.5</priority>
  </url>
  <url>
