+++ next/components/article.tsx
@@ -10,13 +10,11 @@ import { formatDate } from "@/lib/utils";
 import { Article } from "@/lib/zod/article";
 import AnchorNavigation from "./anchorNavigation";
 import avatar from '@/styles/avatar.module.css';
-import addId from "./addId";
 import ArticleLayout from "@/pages/layouts/articlesLayout";
-import { LayoutProps } from "./layout";
 import { ArticleTeaser } from "@/lib/zod/article-teaser";
 import { NewsTeaser } from "@/lib/zod/news-teaser";
-import ArticleSidebar from "./articleSidebar";
 import Link from "next/link";
+import ArticleSidebar from "./articleSidebar";
 
 interface ArticleProps {
   article: Article;
@@ -25,7 +23,6 @@ interface ArticleProps {
 export function Article({ article, ...props }: ArticleProps) {
   const { t } = useTranslation();
   const router = useRouter();
-  const articleBody = addId(article.body?.processed);
 
   return (
     <article {...props}>
@@ -39,9 +36,9 @@ export function Article({ article, ...props }: ArticleProps) {
             title: article.title,
           },
         ]}
-      />
+        />
       {article.field_category && (
-      <span>Category:
+        <span>Category:
         <span className="ml-2 mr-1">
           <Link href={`/all-articles?goToCategory=${article.field_category.name}`} className="text-primary-500">
           {article.field_category.name}
@@ -60,14 +57,14 @@ export function Article({ article, ...props }: ArticleProps) {
           </Link>
           </span> 
         
-          ))}
+        ))}
         </span>
         </div>
       )}
       <HeadingPage>{article.title}</HeadingPage> 
       {article.field_excerpt && (
         <div className="my-4 text-xl">{article.field_excerpt}</div>
-      )}
+        )}
       <div className="mb-4 text-scapaflow flex align-middle gap-2 items-center">
         {article.uid?.field_user_avatar?.uri.url && (
           <Image src={absoluteUrl(article.uid?.field_user_avatar?.uri.url)} width={50} height={50} className={avatar.userAvatar} alt={article.uid?.display_name} />
@@ -90,7 +87,7 @@ export function Article({ article, ...props }: ArticleProps) {
             alt={article.field_image.resourceIdObjMeta.alt}
             className="object-cover mx-auto rounded-lg"
             priority
-          />
+            />
           {article.field_image.resourceIdObjMeta.title && (
             <figcaption className="py-2 text-center text-sm text-scapaflow">
               {article.field_image.resourceIdObjMeta.title}
@@ -99,14 +96,14 @@ export function Article({ article, ...props }: ArticleProps) {
         </figure>
       )}
        {article.field_anchor_nav && (
-          <AnchorNavigation postContent={article.body?.processed} />
-          )}
+         <AnchorNavigation postContent={article.body?.processed} />
+         )}
       {article.body && (
         <FormattedText
-          className="mt-4 text-md/xl text-scapaflow sm:text-lg"
-          html={articleBody}
+        className="mt-4 text-md/xl text-scapaflow sm:text-lg"
+        html={article.body?.processed}
         />
-      )}
+        )}
     </article>
   );
 }
