+++ next/components/formatted-text.tsx
@@ -12,6 +12,8 @@ import parse from "html-react-parser";
 import { isRelative } from "@/lib/utils";
 
 import { env } from "@/env";
+import { add } from "cypress/types/lodash";
+import addId from "@/components/autoID";
 
 const isElement = (domNode: DOMNode): domNode is Element =>
   domNode.type === "tag";
@@ -68,6 +70,10 @@ const options: HTMLReactParserOptions = {
           delete domNode.attribs.value;
         }
 
+        return domNode;
+      }
+      case "h2": {
+        addId(domNode);
         return domNode;
       }
