+++ next/styles/globals.css
@@ -98,15 +98,17 @@
   display: flex;
   flex-wrap: wrap;
   gap: 2rem;
-  margin: 2rem;
-  
+  margin: 1rem;
+  /* width: 100%; */
+  margin: auto
 }
-aside{
-  flex-basis: 30%;
+.main-content-container aside{
+  flex-basis: 25%;
   flex-grow: 1;
 }
-.main-content-container main{
-  flex-basis: 60%;
+.main-content-container article{
+  flex-basis: 70%;
+  flex-grow: 1
 }
 
 h2,h3,h4,h5,h6{
